// src/generated/models/Msfthitl_documentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Msfthitl_documentsmsfthitl_processingstatus = {
  720670000: 'Uploaded',
  720670001: 'Queued',
  720670002: 'Processing',
  720670003: 'Processed',
  720670004: 'Failed'
} as const;
export type Msfthitl_documentsmsfthitl_processingstatus = keyof typeof Msfthitl_documentsmsfthitl_processingstatus;
export const Msfthitl_documentsmsfthitl_reviewstatus = {
  720670000: 'NotRequired',
  720670001: 'PendingReview',
  720670002: 'InReview',
  720670003: 'Approved',
  720670004: 'Rejected'
} as const;
export type Msfthitl_documentsmsfthitl_reviewstatus = keyof typeof Msfthitl_documentsmsfthitl_reviewstatus;
export const Msfthitl_documentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Msfthitl_documentsstatecode = keyof typeof Msfthitl_documentsstatecode;
export const Msfthitl_documentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Msfthitl_documentsstatuscode = keyof typeof Msfthitl_documentsstatuscode;

export interface Msfthitl_documentsBase {
  importsequencenumber?: number;
  msfthitl_documentid: string;
  msfthitl_documentname: string;
  "msfthitl_documenttypeid@odata.bind"?: string;
  msfthitl_extracteddata?: string;
  msfthitl_flaggedforreview?: boolean;
  msfthitl_processedon?: string;
  msfthitl_processingerror?: string;
  msfthitl_processingstatus: Msfthitl_documentsmsfthitl_processingstatus;
  msfthitl_randomdrawvalue?: number;
  msfthitl_reviewcomment?: string;
  msfthitl_reviewedon?: string;
  msfthitl_reviewstatus?: Msfthitl_documentsmsfthitl_reviewstatus;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Msfthitl_documentsstatecode;
  statuscode?: Msfthitl_documentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Msfthitl_documents extends Msfthitl_documentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  msfthitl_documenttypeidname?: string;
  msfthitl_flaggedforreviewname?: string;
  msfthitl_processingstatusname?: string;
  msfthitl_reviewstatusname?: string;
  msfthitl_sourcefile: string;
  msfthitl_sourcefile_name?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  msfthitl_documenttypeid?: object;
  _msfthitl_documenttypeid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Msfthitl_documentsUploadColumnName = 'msfthitl_sourcefile';