// src/generated/models/Msfthitl_documenttypesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Msfthitl_documenttypesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Msfthitl_documenttypesstatecode = keyof typeof Msfthitl_documenttypesstatecode;
export const Msfthitl_documenttypesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Msfthitl_documenttypesstatuscode = keyof typeof Msfthitl_documenttypesstatuscode;

export interface Msfthitl_documenttypesBase {
  importsequencenumber?: number;
  msfthitl_description?: string;
  msfthitl_documenttypeid: string;
  msfthitl_documenttypename: string;
  msfthitl_isactive?: boolean;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Msfthitl_documenttypesstatecode;
  statuscode?: Msfthitl_documenttypesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Msfthitl_documenttypes extends Msfthitl_documenttypesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  msfthitl_isactivename?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}