// src/generated/models/Msfthitl_reviewsettingsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Msfthitl_reviewsettingsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Msfthitl_reviewsettingsstatecode = keyof typeof Msfthitl_reviewsettingsstatecode;
export const Msfthitl_reviewsettingsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Msfthitl_reviewsettingsstatuscode = keyof typeof Msfthitl_reviewsettingsstatuscode;

export interface Msfthitl_reviewsettingsBase {
  importsequencenumber?: number;
  msfthitl_rangemax: number;
  msfthitl_rangemin: number;
  msfthitl_reviewsettingid: string;
  msfthitl_reviewsettingsname: string;
  msfthitl_triggervalue: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Msfthitl_reviewsettingsstatecode;
  statuscode?: Msfthitl_reviewsettingsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Msfthitl_reviewsettings extends Msfthitl_reviewsettingsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}