// src/generated/models/Msfthitl_skillupdaterequestsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Msfthitl_skillupdaterequestsmsfthitl_skillupdatestatus = {
  720670000: 'New',
  720670001: 'InProgress',
  720670002: 'Completed',
  720670003: 'Dismissed'
} as const;
export type Msfthitl_skillupdaterequestsmsfthitl_skillupdatestatus = keyof typeof Msfthitl_skillupdaterequestsmsfthitl_skillupdatestatus;
export const Msfthitl_skillupdaterequestsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Msfthitl_skillupdaterequestsstatecode = keyof typeof Msfthitl_skillupdaterequestsstatecode;
export const Msfthitl_skillupdaterequestsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Msfthitl_skillupdaterequestsstatuscode = keyof typeof Msfthitl_skillupdaterequestsstatuscode;

export interface Msfthitl_skillupdaterequestsBase {
  importsequencenumber?: number;
  "msfthitl_documentid@odata.bind": string;
  msfthitl_documenttypename?: string;
  msfthitl_requestedon?: string;
  msfthitl_resolvedon?: string;
  msfthitl_skillupdaterequestid: string;
  msfthitl_skillupdaterequestname: string;
  msfthitl_skillupdatestatus: Msfthitl_skillupdaterequestsmsfthitl_skillupdatestatus;
  msfthitl_suggestedfix: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Msfthitl_skillupdaterequestsstatecode;
  statuscode?: Msfthitl_skillupdaterequestsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Msfthitl_skillupdaterequests extends Msfthitl_skillupdaterequestsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  msfthitl_documentidname?: string;
  msfthitl_skillupdatestatusname?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  msfthitl_documentid?: object;
  _msfthitl_documentid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}